proceed parmission added

+++ src/Components/ManageItem/ManageItem.js
@@ -14,11 +14,14 @@ const ManageItem = () => {
   }, []);
 
   const deleteItem = (id) => {
-    fetch(`http://localhost:5000/delete/${id}`, {
-      method: "DELETE",
-    });
-    const remaining = products.filter((pd) => pd._id !== id);
-    setProducts(remaining);
+    const proceed = window.confirm("Are You Want to this products!?");
+    if (proceed) {
+      fetch(`http://localhost:5000/delete/${id}`, {
+        method: "DELETE",
+      });
+      const remaining = products.filter((pd) => pd._id !== id);
+      setProducts(remaining);
+    }
   };
   return (
     <Container>
